Añadido de policias en Modulos Riesgo y Recursos

+++ app/Http/Controllers/API/RiskReductionAction/RiskReductionActionController.php
@@ -8,8 +8,10 @@ use App\Http\Requests\RiskReductionAction\StoreRiskReductionActionRequest;
 use App\Http\Requests\RiskReductionAction\UpdateRiskReductionActionRequest;
 use App\Http\Requests\RiskReductionAction\PartialUpdateRiskReductionActionRequest;
 use App\Models\RiskReductionAction\RiskReductionAction;
+use App\Models\RiskFactor\RiskFactor;
 use App\Services\RiskReductionAction\RiskReductionActionService;
 use App\Policies\AccessRiskReductionActionPolicy;
+use App\Policies\AccessRiskFactorPolicy;
 
 class RiskReductionActionController extends Controller
 {
@@ -33,18 +35,18 @@ class RiskReductionActionController extends Controller
 
     public function show(string $id)
     {
-        $action = RiskReductionAction::find($id);
+        $RiskReductionAction = RiskReductionAction::find($id);
 
-        if (!$action) {
+        if (!$RiskReductionAction) {
             return ResponseFormatter::error("Registro no encontrado", 404);
         }
 
-        // if (!(new AccessRiskReductionActionPolicy())->access($action)) {
-        //     return ResponseFormatter::error(
-        //         'Usted no tiene autorización para ver esta acción',
-        //         403
-        //     );
-        // }
+        if (!(new AccessRiskReductionActionPolicy())->access($RiskReductionAction)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para ver esta acción',
+                403
+            );
+        }
 
         $response = $this->service->getById($id);
 
@@ -55,8 +57,43 @@ class RiskReductionActionController extends Controller
         return ResponseFormatter::success($response['message'], $response['code'], $response['data'] ?? []);
     }
 
+    public function getByRiskFactor(string $id)
+    {
+        $RiskFactor = RiskFactor::find($id);
+
+        if (!$RiskFactor) {
+            return ResponseFormatter::error("El factor de riesgo no existe", 404);
+        }
+
+        if (!(new AccessRiskFactorPolicy())->access($RiskFactor)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para ver este factor de riesgo',
+                403
+            );
+        }
+
+        $response = $this->service->getByRiskFactor($id);
+
+        if ($response['error']) {
+            return ResponseFormatter::error($response['message'], $response['code']);
+        }
+
+        return ResponseFormatter::success($response['message'], $response['code'], $response['data'] ?? []);
+    }
+
     public function store(StoreRiskReductionActionRequest $request)
     {
+        
+        $RiskFactor = RiskFactor::find($request->risk_factor_id);
+
+        if (!$RiskFactor) {
+            return ResponseFormatter::error("El factor de riesgo no existe", 404);
+        }
+
+        if (!(new AccessRiskFactorPolicy())->access($RiskFactor)) {
+            return ResponseFormatter::error('No tiene autorización para este plan', 403);
+        }
+
         $data = $request->validated();
         $response = $this->service->create($data);
 
@@ -75,12 +112,12 @@ class RiskReductionActionController extends Controller
             return ResponseFormatter::error("Registro no encontrado", 404);
         }
 
-        // if (!(new AccessRiskReductionActionPolicy())->access($action)) {
-        //     return ResponseFormatter::error(
-        //         'Usted no tiene autorización para modificar esta acción',
-        //         403
-        //     );
-        // }
+        if (!(new AccessRiskReductionActionPolicy())->access($action)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para modificar esta acción',
+                403
+            );
+        }
 
         $response = $this->service->update($request->validated(), $id);
 
@@ -99,12 +136,12 @@ class RiskReductionActionController extends Controller
             return ResponseFormatter::error("Registro no encontrado", 404);
         }
 
-        // if (!(new AccessRiskReductionActionPolicy())->access($action)) {
-        //     return ResponseFormatter::error(
-        //         'Usted no tiene autorización para modificar esta acción',
-        //         403
-        //     );
-        // }
+        if (!(new AccessRiskReductionActionPolicy())->access($action)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para modificar esta acción',
+                403
+            );
+        }
 
         $response = $this->service->partialUpdate($request->validated(), $id);
 
@@ -123,12 +160,12 @@ class RiskReductionActionController extends Controller
             return ResponseFormatter::error("Registro no encontrado", 404);
         }
 
-        // if (!(new AccessRiskReductionActionPolicy())->access($action)) {
-        //     return ResponseFormatter::error(
-        //         'Usted no tiene autorización para eliminar esta acción',
-        //         403
-        //     );
-        // }
+        if (!(new AccessRiskReductionActionPolicy())->access($action)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para eliminar esta acción',
+                403
+            );
+        }
 
         $response = $this->service->delete($id);
